tiranga login
Tiranga Game registration can be completed very easily. Just follow our step-by-step procedure for Tiranga Registration.

Head to the Tiranga Game Website or open the app.
On the homepage, you need to click on the Tiranga Registration.
Enter your basic details like name, mobile number, and email ID in the required column.
After entering the details, the OTP will be sent to the registered mobile number.
Enter the OTP, and your registration process for the Tiranga game will be completed.